Also test against Python 3.7 on Travis CI
This requires splitting the build into jobs which can individually
specify which Ubuntu version they run on. This is because Python 3.7 is
only available on Xenial hosts on Travis CI, but Python 3.4 is no longer
available on Xenial.

I've also changed the Travis configuration to install flake8 during the
install step, rather than the script step. Now if flake8 fails to
install, the test will show as errored rather than failed.