Reorder arguments in ServerClient.apply_environment
PR #66 added ServerClient.apply_environment with the arguments
(environment_id, server_id), however we usually put ids in the
order (project, server, others). In this specific case, the server ID
seems more 'primary' as the server is the main resource from the point
of view of the client.

This method does appear to yet be used by the CLI so I propose merging
this PR before https://github.com/facultyai/faculty-cli/pull/11/ is
merged.