Change interface for create() and update()
The create() and update() methods are rewritten
to expect the name, specification and optional
description of an environment to be passed explicitly;
an EnvironmentUpdateCreate object is then created internally
instead of being passed to these functions by the user.
In addition, extra tests are written for the case when
the decription is None.